Refactorizes some classes and packages
