+++ core/src/test/scala/query/mapper/ByteBufferMapperSpec.scala
@@ -15,13 +15,12 @@
  */
 
 package freestyle.cassandra
-package mapper
+package query.mapper
 
 import java.nio.ByteBuffer
 
 import cats.MonadError
 import com.datastax.driver.core.{ProtocolVersion, TypeCodec}
-import freestyle.cassandra.mapper.FieldToByteBuffer._
 import org.scalacheck.Prop._
 import org.scalatest.{Matchers, WordSpec}
 import org.scalacheck.ScalacheckShapeless._
